Create a place to keep and test modifications to how we want projects displayed
+ Over time we can move project display modifications into here, where we can test them in isolation from the component framework.
